requirements.txt
aiohttp==2.1.0
async-timeout==1.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==0.12.2
gevent==1.3.4
greenlet==0.4.13
idna==2.7
itsdangerous==0.24
Jinja2==2.9.6
locustio==0.8.1
MarkupSafe==1.0
matplotlib==2.0.2
msgpack-python==0.5.6
multidict==2.1.6
numpy==1.14.5
pandas==0.23.2
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.0.0
requests==2.19.1
scikit-learn==0.19.1
scipy==0.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.12.2
wincertstore==0.2
yarl==0.10.3
